requirements.txt
vosk==0.3.45
edge-tts==7.2.3
playsound3==2.2.2
sounddevice==0.5.1
imageio-ffmpeg==0.4.9
requests==2.31.0
psutil==5.9.5
numpy>=1.24
python-dotenv
rich
google-genai
google